Update README.md
Emily in Paris follows Emily Cooper, an American college graduate in her late twenties with a Master's degree in communication who is from Chicago; she moves to Paris for an unexpected job opportunity. She is tasked with bringing an American point of view and social media presence to a venerable French marketing firm.